# Try to optimize Optimization a bit