--- a/src/test/scala/mjis/CompilerTestMatchers.scala
+++ b/src/test/scala/mjis/CompilerTestMatchers.scala
@@ -169,7 +169,6 @@ trait CompilerTestMatchers {
 
       val opt = new Optimizer((), Config())
       opt.generalOptimizations = opt.generalOptimizations.filter(!excludedOptimizations.contains(_))
-      opt.highLevelOptimizations = opt.highLevelOptimizations.filter(!excludedOptimizations.contains(_))
 
       opt.result
       opt.dumpResult(null)